refactor: refactor file system & replace defaults
