+++ config/siteMetadata.js
@@ -1,0 +1,5 @@
+module.exports = {
+  title: `FCC MCR`,
+  description: `Home of the Official FreeCodeCamp Manchester Group`,
+  author: `FCC Manchester`,
+}
