+++ src/components/Layout/Header/index.js
@@ -5,7 +5,7 @@ import React from "react"
 const Header = ({ siteTitle }) => (
   <header
     style={{
-      background: `rebeccapurple`,
+      background: `#4D844F`,
       marginBottom: `1.45rem`,
     }}
   >
